Moved working files to a src folder
