+++ src/plugin_manager.py
