+++ src/plugins/doge.py
