+++ src/plugins/manager.py
