+++ src/plugins/stats.py
